azure_functions/requirements.txt
# --- Azure Functions runtime ---
azure-functions
# --- Azure SDKs ---
azure-storage-blob
azure-cosmos
azure-search-documents
azure-identity
azure-core
# --- LangChain ---
langchain
langchain-core
langchain-community
langchain-text-splitters

# --- Utilidades ---
python-dotenv
requests